+++ test/common/http/BUILD
@@ -193,54 +193,30 @@ envoy_cc_fuzz_test(
 
 envoy_cc_test(
     name = "conn_manager_impl_test",
-    srcs = ["conn_manager_impl_test.cc"],
+    srcs = [
+        # Split to avoid compiler OOM, especially on ASAN.
+        "conn_manager_impl_test.cc",
+        "conn_manager_impl_test_2.cc",
+        "conn_manager_impl_test_base.cc",
+        "conn_manager_impl_test_base.h",
+    ],
     shard_count = 3,
     deps = [
-        "//include/envoy/access_log:access_log_interface",
-        "//include/envoy/buffer:buffer_interface",
-        "//include/envoy/event:dispatcher_interface",
-        "//include/envoy/http:request_id_extension_interface",
-        "//include/envoy/tracing:http_tracer_interface",
-        "//source/common/access_log:access_log_lib",
-        "//source/common/buffer:buffer_lib",
-        "//source/common/common:macros",
-        "//source/common/event:dispatcher_lib",
-        "//source/common/formatter:substitution_formatter_lib",
         "//source/common/http:conn_manager_lib",
         "//source/common/http:context_lib",
-        "//source/common/http:date_provider_lib",
-        "//source/common/http:exception_lib",
-        "//source/common/http:header_map_lib",
-        "//source/common/http:headers_lib",
-        "//source/common/http:request_id_extension_lib",
-        "//source/common/network:address_lib",
-        "//source/common/stats:stats_lib",
-        "//source/common/upstream:upstream_includes",
-        "//source/common/upstream:upstream_lib",
         "//source/extensions/access_loggers/file:file_access_log_lib",
-        "//source/server:drain_manager_lib",
-        "//test/mocks:common_lib",
         "//test/mocks/access_log:access_log_mocks",
-        "//test/mocks/buffer:buffer_mocks",
+        "//test/mocks/event:event_mocks",
         "//test/mocks/http:http_mocks",
         "//test/mocks/local_info:local_info_mocks",
         "//test/mocks/network:network_mocks",
         "//test/mocks/router:router_mocks",
         "//test/mocks/runtime:runtime_mocks",
         "//test/mocks/server:factory_context_mocks",
-        "//test/mocks/server:instance_mocks",
-        "//test/mocks/server:overload_manager_mocks",
         "//test/mocks/ssl:ssl_mocks",
-        "//test/mocks/tracing:tracing_mocks",
-        "//test/mocks/upstream:cluster_manager_mocks",
-        "//test/mocks/upstream:host_mocks",
-        "//test/mocks/upstream:thread_local_cluster_mocks",
         "//test/test_common:logging_lib",
+        "//test/test_common:simulated_time_system_lib",
         "//test/test_common:test_runtime_lib",
-        "//test/test_common:test_time_lib",
-        "@envoy_api//envoy/extensions/filters/network/http_connection_manager/v3:pkg_cc_proto",
-        "@envoy_api//envoy/type/tracing/v3:pkg_cc_proto",
-        "@envoy_api//envoy/type/v3:pkg_cc_proto",
     ],
 )
